fix: rename use-my-account-v2-email.ts.ts to .ts and fix import paths
