+++ feature-libs/user/profile/components/update-email/index.ts
@@ -7,5 +7,5 @@
 export * from './update-email-component.service';
 export * from './update-email.component';
 export * from './update-email.module';
-export * from './use-my-account-v2-email.ts';
+export * from './use-my-account-v2-email';
 export * from './my-account-v2-email.component';
